requirements.txt
google-genai>=0.8.0
google-auth>=2.0.0
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.0
pyarrow>=10.0.0
xlrd>=2.0.0
